odb: move gc into a separate function
