+++ dvc/objects/db.py
@@ -356,46 +356,6 @@ class ObjectDB:
     def _remove_unpacked_dir(self, hash_):
         pass
 
-    def gc(self, used, jobs=None, cache_odb=None, shallow=True):
-        from dvc.data.tree import Tree
-
-        if self.read_only:
-            raise ObjectDBPermissionError("Cannot gc read-only ODB")
-        if not cache_odb:
-            cache_odb = self
-        used_hashes = set()
-        for hash_info in used:
-            used_hashes.add(hash_info.value)
-            if hash_info.isdir and not shallow:
-                tree = Tree.load(cache_odb, hash_info)
-                used_hashes.update(
-                    entry_obj.hash_info.value for _, entry_obj in tree
-                )
-
-        def _is_dir_hash(_hash):
-            from dvc.hash_info import HASH_DIR_SUFFIX
-
-            return _hash.endswith(HASH_DIR_SUFFIX)
-
-        removed = False
-        # hashes must be sorted to ensure we always remove .dir files first
-        for hash_ in sorted(
-            self.all(jobs, self.fs_path),
-            key=_is_dir_hash,
-            reverse=True,
-        ):
-            if hash_ in used_hashes:
-                continue
-            fs_path = self.hash_to_path(hash_)
-            if _is_dir_hash(hash_):
-                # backward compatibility
-                # pylint: disable=protected-access
-                self._remove_unpacked_dir(hash_)
-            self.fs.remove(fs_path)
-            removed = True
-
-        return removed
-
     def list_hashes_exists(self, hashes, jobs=None, name=None):
         """Return list of the specified hashes which exist in this fs.
         Hashes will be queried individually.
